Move shareable logging code into common/commercial
